requirements.txt
Django>=3.0,<4.0
djangorestframework==3.13.1
drf-yasg==1.20.0
python-decouple==3.6
psycopg2>=2.8
gunicorn==19.9.0
redis==3.5.3
celery==5.1.2
django-reversion==3.0.8
pytest==6.0.1
djangorestframework-simplejwt==4.4.0